extracted language setting in specialized hook
